+++ packages/swagger-editor/src/components/SwaggerEditor/index.ts
@@ -1,2 +1,1 @@
 export { default as SwaggerEditor } from './SwaggerEditor.vue'
-export { default as SwaggerEditorGettingStarted } from './SwaggerEditorGettingStarted.vue'