+++ packages/hono-api-reference/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @scalar/hono-api-reference
 
+## 0.2.0
+
+### Minor Changes
+
+- f38bb61: chore: remove swagger editor tabs
+
+### Patch Changes
+
+- 349589a: feat: expose spec content updates
+- Updated dependencies [349589a]
+- Updated dependencies [f38bb61]
+  - @scalar/api-reference@1.5.0
+
+## 0.1.22
+
+### Patch Changes
+
+- 95fbc01: fix: hono middleware fails to read configuration.spec.content
+- Updated dependencies [b65840b]
+- Updated dependencies [95fbc01]
+- Updated dependencies [b65840b]
+- Updated dependencies [b974eeb]
+  - @scalar/api-reference@1.4.0
+
 ## 0.1.21
 
 ### Patch Changes